requirements.txt
discord.py==2.0.0
aiohttp==3.8.1
python-dotenv==0.19.2